+++ Source/UnityAdsAdapter.swift
@@ -141,7 +141,12 @@ final class UnityAdsAdapter: NSObject, PartnerAdapter {
         case .banner:
             return try UnityAdsAdapterBannerAd(adapter: self, request: request, delegate: delegate)
         default:
-            throw error(.loadFailureUnsupportedAdFormat)
+            // Not using the `.adaptiveBanner` case directly to maintain backward compatibility with Chartboost Mediation 4.0
+            if request.format.rawValue == "adaptive_banner" {
+                return try UnityAdsAdapterBannerAd(adapter: self, request: request, delegate: delegate)
+            } else {
+                throw error(.loadFailureUnsupportedAdFormat)
+            }
         }
     }